Add key pages and navigation
Gave the navigation some pizazz. Whenever a button is pressed it will
become a raised button and render all of the other buttons flat. Along
with this i added the other pages that will be included in the site.
Goind to consider changing the gradiant to a per page thing.